requirements.txt
asgiref==3.1.4
asn1crypto==0.24.0
attrs==19.1.0
autobahn==19.7.1
Automat==0.7.0
cffi==1.12.3
channels==2.2.0
constantly==15.1.0
cryptography==2.7
daphne==2.3.0
Django==2.2.3
django-cors-headers==3.0.2
djangorestframework==3.10.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
pycparser==2.19
PyHamcrest==1.9.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
Twisted==19.2.1
txaio==18.8.1
zope.interface==4.6.0
